# packages/epist/epist-1.2.2.tar.gz/epist-1.2.2/openapi_v2.json
{"openapi":"3.1.0","info":{"title":"Audio RAG Platform","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/audio/transcribe_url":{"post":{"tags":["Audio"],"summary":"Transcribe Url","description":"Transcribe audio from a URL.","operationId":"transcribe_url_api_v1_audio_transcribe_url_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AudioUrlRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AudioResourceRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audio/upload":{"post":{"tags":["Audio"],"summary":"Upload Audio","description":"Upload an audio file to the platform.\n\n- **file**: The audio file (MP3, WAV, M4A, etc.)","operationId":"upload_audio_api_v1_audio_upload_post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_audio_api_v1_audio_upload_post"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AudioResourceRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audio/{audio_id}":{"get":{"tags":["Audio"],"summary":"Get Audio","description":"Get an audio resource by ID.","operationId":"get_audio_api_v1_audio__audio_id__get","parameters":[{"name":"audio_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Audio Id"}},{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AudioResourceRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/audio/{audio_id}/transcript":{"get":{"tags":["Audio"],"summary":"Get Transcript","description":"Get the transcript for an audio resource, including segments.","operationId":"get_transcript_api_v1_audio__audio_id__transcript_get","parameters":[{"name":"audio_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Audio Id"}},{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TranscriptReadWithSegments"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/search/":{"post":{"tags":["Search"],"summary":"Search","description":"Perform Hybrid Search (Vector + Keyword) on transcripts.","operationId":"search_api_v1_search__post","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SearchResponseItem"},"title":"Response Search Api V1 Search  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/stats/":{"get":{"tags":["Stats"],"summary":"Get Stats","description":"Get system statistics: counts of audio files, transcripts, and segments.","operationId":"get_stats_api_v1_stats__get","parameters":[{"name":"x-api-key","in":"header","required":false,"schema":{"type":"string","title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/tasks/process-audio":{"post":{"tags":["Tasks"],"summary":"Process Audio Task","description":"Handler for Cloud Tasks to process audio.\nVerifies OIDC token to ensure request is authorized.","operationId":"process_audio_task_api_v1_tasks_process_audio_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessAudioTask"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/api-keys":{"post":{"tags":["Auth"],"summary":"Create Api Key","description":"Create a new API key for the authenticated user.","operationId":"create_api_key_api_v1_auth_api_keys_post","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Create Api Key Api V1 Auth Api Keys Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Auth"],"summary":"List Api Keys","description":"List all active API keys for the authenticated user.","operationId":"list_api_keys_api_v1_auth_api_keys_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true},"title":"Response List Api Keys Api V1 Auth Api Keys Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/api-keys/{key_id}":{"delete":{"tags":["Auth"],"summary":"Revoke Api Key","description":"Revoke an API key.","operationId":"revoke_api_key_api_v1_auth_api_keys__key_id__delete","parameters":[{"name":"key_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Key Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Auth"],"summary":"Rename Api Key","description":"Rename an API key.","operationId":"rename_api_key_api_v1_auth_api_keys__key_id__put","parameters":[{"name":"key_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Key Id"}},{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Rename Api Key Api V1 Auth Api Keys  Key Id  Put"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/logs/":{"get":{"tags":["Logs"],"summary":"Get Logs","operationId":"get_logs_api_v1_logs__get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"status_code","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Status Code"}},{"name":"method","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Method"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RequestLog"},"title":"Response Get Logs Api V1 Logs  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/completions":{"post":{"tags":["Chat"],"summary":"Chat Completions","description":"Chat with your audio knowledge base.","operationId":"chat_completions_api_v1_chat_completions_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/traces/":{"get":{"tags":["Traces"],"summary":"List Traces","description":"List trace events.","operationId":"list_traces_api_v1_traces__get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":50,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"trace_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Trace Id"}},{"name":"component","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Component"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TraceEvent"},"title":"Response List Traces Api V1 Traces  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/traces/{trace_id}":{"get":{"tags":["Traces"],"summary":"Get Trace Details","description":"Get all events for a specific trace ID (waterfall).","operationId":"get_trace_details_api_v1_traces__trace_id__get","parameters":[{"name":"trace_id","in":"path","required":true,"schema":{"type":"string","title":"Trace Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}},{"name":"x-api-key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Api-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TraceEvent"},"title":"Response Get Trace Details Api V1 Traces  Trace Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AudioResourceRead":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"source_url":{"type":"string","title":"Source Url"},"duration_seconds":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Duration Seconds"},"status":{"type":"string","title":"Status","default":"pending"},"meta_data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Meta Data"},"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["source_url","id","created_at"],"title":"AudioResourceRead"},"AudioUrlRequest":{"properties":{"audio_url":{"type":"string","title":"Audio Url"},"rag_enabled":{"type":"boolean","title":"Rag Enabled","default":true},"language":{"type":"string","title":"Language","default":"en"}},"type":"object","required":["audio_url"],"title":"AudioUrlRequest"},"Body_upload_audio_api_v1_audio_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File","description":"Audio file to upload"}},"type":"object","required":["file"],"title":"Body_upload_audio_api_v1_audio_upload_post"},"ChatMessage":{"properties":{"role":{"type":"string","title":"Role"},"content":{"type":"string","title":"Content"}},"type":"object","required":["role","content"],"title":"ChatMessage"},"ChatRequest":{"properties":{"messages":{"items":{"$ref":"#/components/schemas/ChatMessage"},"type":"array","title":"Messages"},"stream":{"type":"boolean","title":"Stream","default":false},"model":{"type":"string","title":"Model","default":"gpt-3.5-turbo"}},"type":"object","required":["messages"],"title":"ChatRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ProcessAudioTask":{"properties":{"audio_id":{"type":"string","format":"uuid","title":"Audio Id"},"audio_url":{"type":"string","title":"Audio Url"}},"type":"object","required":["audio_id","audio_url"],"title":"ProcessAudioTask"},"RequestLog":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"request_id":{"type":"string","title":"Request Id"},"method":{"type":"string","title":"Method"},"path":{"type":"string","title":"Path"},"status_code":{"type":"integer","title":"Status Code"},"latency_ms":{"type":"number","title":"Latency Ms"},"ip_address":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ip Address"},"user_agent":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"User Agent"},"api_key_id":{"anyOf":[{"type":"string","format":"uuid"},{"type":"null"}],"title":"Api Key Id"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["request_id","method","path","status_code","latency_ms"],"title":"RequestLog"},"SearchRequest":{"properties":{"query":{"type":"string","title":"Query"},"limit":{"type":"integer","title":"Limit","default":10}},"type":"object","required":["query"],"title":"SearchRequest"},"SearchResponseItem":{"properties":{"id":{"title":"Id"},"text":{"type":"string","title":"Text"},"start":{"type":"number","title":"Start"},"end":{"type":"number","title":"End"},"score":{"type":"number","title":"Score"},"methods":{"items":{"type":"string"},"type":"array","title":"Methods"}},"type":"object","required":["id","text","start","end","score","methods"],"title":"SearchResponseItem"},"TraceEvent":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"trace_id":{"type":"string","title":"Trace Id"},"span_id":{"type":"string","title":"Span Id"},"parent_span_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Parent Span Id"},"event_type":{"type":"string","title":"Event Type"},"component":{"type":"string","title":"Component"},"name":{"type":"string","title":"Name"},"inputs":{"additionalProperties":true,"type":"object","title":"Inputs","default":{}},"outputs":{"additionalProperties":true,"type":"object","title":"Outputs","default":{}},"meta":{"additionalProperties":true,"type":"object","title":"Meta"},"start_time":{"type":"string","format":"date-time","title":"Start Time"},"end_time":{"type":"string","format":"date-time","title":"End Time"},"latency_ms":{"type":"number","title":"Latency Ms"},"status":{"type":"string","title":"Status","default":"success"},"error_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error Message"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["trace_id","span_id","event_type","component","name","start_time","end_time","latency_ms"],"title":"TraceEvent"},"TranscriptReadWithSegments":{"properties":{"audio_resource_id":{"type":"string","format":"uuid","title":"Audio Resource Id"},"language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Language"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"text":{"type":"string","title":"Text"},"id":{"type":"string","format":"uuid","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"segments":{"items":{"$ref":"#/components/schemas/TranscriptSegmentRead"},"type":"array","title":"Segments"}},"type":"object","required":["audio_resource_id","text","id","created_at","segments"],"title":"TranscriptReadWithSegments"},"TranscriptSegmentRead":{"properties":{"start":{"type":"number","title":"Start"},"end":{"type":"number","title":"End"},"text":{"type":"string","title":"Text"},"speaker":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Speaker"},"confidence":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Confidence"},"words":{"anyOf":[{"items":{"additionalProperties":true,"type":"object"},"type":"array"},{"type":"null"}],"title":"Words"},"embedding":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Embedding"},"content_vector":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content Vector"},"id":{"type":"string","format":"uuid","title":"Id"},"transcript_id":{"type":"string","format":"uuid","title":"Transcript Id"}},"type":"object","required":["start","end","text","id","transcript_id"],"title":"TranscriptSegmentRead"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}